changes based on first impression
