+++ trunk/android/StarLib/src/edu/illinois/mitra/starl/gvh/RealGlobalVarHolder.java
@@ -4,10 +4,9 @@ import android.content.Context;
 import android.os.Handler;
 
 import java.util.Map;
-import java.util.Vector;
 
+import edu.illinois.mitra.starl.comms.DummyGPS;
 import edu.illinois.mitra.starl.comms.SmartUdpComThread;
-import edu.illinois.mitra.starl.comms.UdpGpsReceiver;
 import edu.illinois.mitra.starl.interfaces.TrackedRobot;
 import edu.illinois.mitra.starl.models.ModelARDrone2;
 import edu.illinois.mitra.starl.models.Model_iRobot;
@@ -18,8 +17,6 @@ import edu.illinois.mitra.starl.motion.MotionAutomation_ARDrone2;
 import edu.illinois.mitra.starl.motion.MotionAutomatonMiniDrone;
 import edu.illinois.mitra.starl.motion.MotionAutomaton_iRobot;
 import edu.illinois.mitra.starl.motion.ReachAvoid;
-import edu.illinois.mitra.starl.objects.ObstacleList;
-import edu.illinois.mitra.starl.objects.PositionList;
 
 /**
  * Extension of the GlobalVarHolder class for use in physical implementations of StarL applications
@@ -43,9 +40,9 @@ public class RealGlobalVarHolder extends GlobalVarHolder {
 		super.trace = new Trace(name, "/sdcard/trace/", this);
 		super.plat = new RealAndroidPlatform(handler);
 		super.comms = new Comms(this, new SmartUdpComThread(this));
-		super.gps = new Gps(this, new UdpGpsReceiver(this,"192.168.1.100",4000,new PositionList(),
-							new PositionList(), new ObstacleList(), new Vector<ObstacleList>(3,2))  );
-//		super.gps = new Gps(this, new DummyGPS(this) );
+//		super.gps = new Gps(this, new UdpGpsReceiver(this,"192.168.1.100",4000,new PositionList(),
+//							new PositionList(), new ObstacleList(), new Vector<ObstacleList>(3,2))  );
+		super.gps = new Gps(this, new DummyGPS(this) );
 		plat.model = initpos;
 		plat.reachAvoid = new ReachAvoid(this);
         if(initpos instanceof Model_iRobot) {
