+++ trunk/android/StarLib/src/edu/illinois/mitra/starl/models/ModelARDrone2.java
@@ -2,11 +2,6 @@ package edu.illinois.mitra.starl.models;
 
 import android.util.Log;
 
-import de.yadrone.base.ARDrone;
-import de.yadrone.base.IARDrone;
-import de.yadrone.base.command.CommandManager;
-import de.yadrone.base.command.LEDAnimation;
-import de.yadrone.base.navdata.NavDataManager;
 import edu.illinois.mitra.starl.exceptions.ItemFormattingException;
 import edu.illinois.mitra.starl.interfaces.TrackedRobot;
 import edu.illinois.mitra.starl.objects.ItemPosition;
@@ -18,37 +13,16 @@ import edu.illinois.mitra.starl.objects.PositionList;
  * Created by SC on 6/6/16.
  */
 public class ModelARDrone2 extends ItemPosition implements TrackedRobot{
-    class ARDrone2_AttitudeListn implements de.yadrone.base.navdata.AttitudeListener{
-        private String tag;
-
-        ARDrone2_AttitudeListn(String name){
-            tag = name;
-        }
-
-        public void attitudeUpdated(float pitch, float roll, float yaw)
-        {
-            //Log.i(tag + " rcvData", "Pitch: " + pitch + " Roll: " + roll + " Yaw: " + yaw);
-        }
-
-        public void attitudeUpdated(float pitch, float roll) { ;}
-        public void windCompensation(float pitch, float roll) { ;}
-    }
-
     public String ipAddr;
-    public IARDrone droneInstance;
-    public CommandManager cmd;//TODO: re-encapsulate this thing
-    private NavDataManager nav;
     //the position/velocity/status information
     //angular velocity in radius/second, regular velocity in meter/second
     //position x,y,z; name defined in ItemPosition
     public double currYaw, currPitch, currRoll;
     public double vYaw, vPitch, vRoll;
     public double vX, vY, vZ;
-    public int maxSpeed;//in percentage
 
     //configurations supported by the drone
     //TODO: public double maxAltitude, minAltitude;
-    public boolean outdoor;
 
     public ModelARDrone2(String descStr) throws ItemFormattingException {
         String[] parts = descStr.replace(",", "").split("\\|");
@@ -67,62 +41,6 @@ public class ModelARDrone2 extends ItemPosition implements TrackedRobot{
         initHelper();
     }
 
-    public ModelARDrone2(String descStr, ModelARDrone2 old) throws ItemFormattingException {
-        String[] parts = descStr.replace(",", "").split("\\|");
-        if(parts.length==10){
-            this.name = parts[1];
-            this.x = Integer.parseInt(parts[2]);
-            this.y = Integer.parseInt(parts[3]);
-            this.z = Integer.parseInt(parts[4]);
-            this.currYaw = Integer.parseInt(parts[5]);
-            this.currPitch = Integer.parseInt(parts[6]);
-            this.currRoll = Integer.parseInt(parts[7]);
-            this.ipAddr = parts[8];
-        } else {
-            throw new ItemFormattingException("Should be length 10, is length " + parts.length);
-        }
-        CopySetupDrone(old);
-    }
-
-    private void CopyFull(ModelARDrone2 old){
-        this.ipAddr = old.ipAddr;
-        this.droneInstance = old.droneInstance;
-        this.cmd = old.cmd;
-        this.nav = old.nav;
-        this.currPitch = old.currPitch;
-        this.currRoll = old.currRoll;
-        this.currYaw = old.currYaw;
-        this.vYaw = old.vYaw;
-        this.vPitch = old.vPitch;
-        this.vRoll = old.vRoll;
-        this.vX = old.vX;
-        this.vY = old.vY;
-        this.vZ = old.vZ;
-        this.maxSpeed = old.maxSpeed;
-        this.outdoor = old.outdoor;
-        this.x = old.x;
-        this.y = old.y;
-        this.z = old.z;
-    }
-
-    private void CopySetupDrone(ModelARDrone2 old) throws ItemFormattingException{
-        if(old.ipAddr == null || old.droneInstance == null || old.cmd == null) {
-            String out = "The OLD drone does not seems to be properly set up.";
-            if(old.ipAddr==null) out+=" no IPADDR.";
-            if(old.droneInstance==null) out+=" no DR_INSTANCE.";
-            if(old.cmd==null) out+=" no CMD.";
-            throw new ItemFormattingException(out);
-        }
-        this.ipAddr = old.ipAddr;
-        this.droneInstance = old.droneInstance;
-        this.cmd = old.cmd;
-        this.nav = old.nav;
-        this.vX = old.vX;
-        this.vY = old.vY;
-        this.vZ = old.vZ;
-        this.maxSpeed = old.maxSpeed;
-        this.outdoor = old.outdoor;
-    }
 
     public ModelARDrone2(String name, int x, int y){
         super(name, x, y, 0);
@@ -143,33 +61,17 @@ public class ModelARDrone2 extends ItemPosition implements TrackedRobot{
     private void initHelper(){
         if(this.ipAddr==null)
             this.ipAddr = "192.168.1.10";
-        droneInstance = new ARDrone(ipAddr, null);
         vX=0;
         vY=0;
         vZ=0;
         vYaw=0;
         vPitch=0;
         vRoll=0;
-        outdoor = false;
-        maxSpeed = 5;
     }
 
     @Override
     public void initialize(){
-        try{
-            Log.i("Model ARDrone2", "initialization called, this.ip="+this.ipAddr);
-            droneInstance.reset();
-            droneInstance.start();
-            cmd = droneInstance.getCommandManager();
-            nav = droneInstance.getNavDataManager();
-            nav.addAttitudeListener(new ARDrone2_AttitudeListn(this.name));
-            droneInstance.setSpeed(maxSpeed);
-            cmd.setOutdoor(outdoor, outdoor);
-            cmd.setLedsAnimation(LEDAnimation.BLINK_ORANGE, 3, 10);//some sig for us
-        }catch (Exception exc)
-        {
-            exc.printStackTrace();
-        }
+        Log.i("Model ARDrone2", "initialization called, this.ip="+this.ipAddr);
     }
 
     //TODO, temp stub here
